ib_core: Use weak ordering for data registered memory
Some architectures support weak ordering in which case better
performance is possible. IB registered memory used for data can be
weakly ordered because the completion queues' buffers are
registered as strongly ordered. This will result in flushing all data
related outstanding DMA requests by the HCA when a completion is DMAed
to a completion queue buffer.
This patch will allow weak ordering for data if ib_core is loaded with
the module parameter, allow_weak_ordering, set to a none zero value.

Signed-off-by: Eli Cohen <eli@mellanox.co.il>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Vladimir Sokolovsky <vlad@mellanox.com>