mlx4: Increase the command timeout to 60 seconds [topic: increase_command_timeout]
On some environments we've noticed that 10 second timeout for FW
commands is just not enough. For instance, lets take a look on the
SW2HW_CQ command, where the FW handles a CQ buffer initialization
which could hold up to 256MB (depends on the CQ and CQE sizes). Over
PCIe gen1 slot we can reach about 2GB/s memory BW and when the HW is
congested with network traffic it handles CQE writes in lower rate of
~200MB/s so this whole process may take up to ~1.3s for executing a
single command. Taking into account that up to 16 commands can pend for
FW handling at the same time, the whole process might take ~20 seconds.
Since we want to be resilient of future FW timeouts for non-error flows,
we took a high enough margins to cover such cases.

Issue: 430319
Issue: 428046
Change-Id: I5ed0a7fe662333a30aeabc12179ec60bb0b4d335
Signed-off-by: Eyal Perry <eyalpe@mellanox.com>