IB/ipoib: Remove datagram mode check for ipoib flow steering
When ipoib UD qp is created, the user configuration (datagram/connected)
Is still unknown to the driver.
Because of that the ipoib UD qp will be created from the HW steering range in both configurations.
The user has to be aware, ipoib flow steering functionality won't work in connected mode, only in datagram.

Signed-off-by: Hadar Hen Zion <hadarh@mellanox.com>

Change-Id: I7947b01cd223770f0c64907e93258b6d005a30e1