mlx4_core: Add mechanism for non trackable mem free
Fixes RM #92081

Bug description:
The function ib_alloc_device allocates memory for the device that
eventually released by the kernel, therefore this allocation is not trackable
by memtrack.
The problem rises in case we can't initialize the device and we need to
deallocate it ourselves.
Under this scenario memtrack raises an error since it is not handling the mem
allocation but do handle the mem free.

This patch introduces 2 changes.

First change:
Since the is_skb_allocation func is looking for all non trackable allocation,
I replaced the is_skb_allocation function name to is_non_trackable_alloc_func.
The function is also changed to work with tables instead of long "OR" statement.

Second change:
To resolve the bug I added function is_non_trackable_free_func which defines
the cases in which we free an untrackable memory. This function allows memtrack
skip the tracking of such memory free.

V2:
   Adjusted for mlnx_ofed_next branch.

Signed-off-by: Moshe Lazer <moshel@mellanox.co.il>
Reviewed-by: Eli Cohen <eli@mellanox.co.il>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>