mlx4_core: fix broken logic when removing promisc qp
This commit fixes the bug that was introduced by:
    commit 5a727a2166c0cb806f60700ceab8757d01988204
    Author: Dotan Barak <dotanb@dev.mellanox.co.il>
    Date:   Wed Aug 1 11:50:21 2012 +0300

           net/mlx4_core: make sure that negative array index isn't used

If doing the break when location found i is not incremented up to members count
and wrong MGM QP index is filled with 0.
Bug was found during NCSI testing. It caused to NCSI QP to be overwritten.
From now use members count to swap the last QP with removed one.
Fix also the second occurence of this flow in mlx4_qp_detach_common().

issue: 101276

Change-Id: I88988c82776491d9bcb862a1be1abe035f015e51
Signed-off-by: Eugenia Emantayev <eugenia@mellanox.co.il>
Reviewed-by: Alexander Guller <alexg@mellanox.com>
Signed-off-by: Eugenia Emantayev <eugenia@mellanox.co.il>