mlx4_en: Data path structures are allocated per numa node.
For each RX/TX ring, and its CQ, allocation is done on a numa node
that corresponds to the core that the data structure should operate
on.
The assumption is that the core number is reflected by the ring index.
(IRQ affinity hints will be added).
The affected allocations are the ring/CQ data structures, the TX/RX info
and the shared HW/SW buffer.
For TX rings, each core has rings of all UPs

Signed-off-by: Yevgeny Petrilin <yevgenyp@mellanox.co.il>