mlx4_en: Fixing race when closing TX queue
There is a possible race where the TX completion handler
can clean the entire TX queue between the decission that
the queue is full and actually closing it.
To avoid this situation checking again the the queue is really full,
if not, reopen the transmit and continue with sending the packet.
In addition remove ring->blocked flag.

V2: Adjusted for kernel 3.7-rc4
blocked flag was already removed, in commit c18520bd.

Signed-off-by: Yevgeny Petrilin <yevgenyp@mellanox.co.il>
Signed-off-by: Eugenia Emantayev <eugenia@mellanox.co.il>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>