mlx4_ib: Adapt driver to changes in steering range API changes in firmware
To improve performance when no UC steering rule is set (when there is no need
to use the replication rules in MCG table).
Also, TSS QP numbers must be adjacent to the parent QP number who need to be in
the steering QP range. Fixed driver to act accordingly.

V2: Adjusted for RoCE/SRIOV patches. Also did some indentation cleanup
and cleaned up order of testing conditions in "if" statements.

Change-Id: I11010171808047152234d447e31aa478b3708e78
Signed-off-by: Eitan Rabin <rabin@mellanox.com>
Signed-off-by: Amir Vadai <amirv@mellanox.com>
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>