ib/ipoib: fix kernel panic when restarting with vma  flag
Issue: 33890

Issue title
Kernel panic when restarting the driver after MLNX_OFED installation using '--vma' flag

It was assumed that ipoib_get_stats (.ndo_get_stats) is called under rtnl_lock.
This is true only if "ip link show <interface> command is used. All other access
methods are not protected. For example "ifconfig <interfcae>" or
"cat /sys/class/net/<interface>/statistics/*"
Also the kernel periodically calls the ".ndo_get_stats" without protection.
The access to this routine need to be controlled as it polls the Rx and Tx rings
which may be recreated as done when vma is used.

Signed-off-by: Shlomo Pongratz <shlomop@mellanox.com>