ipoib: fix a deadlock at "carrier on task"
Upon a response to a broadcast group join request, when IPoIB module is loaded and immediately unloaded,
the machine might enter a deadlock situation.

We use a new private lock instead of rtnl lock to synchronize between ipoib_stop and carrier on task.

Using the rtnl lock at "carrier on task" may cause a deadlock.
This lock was taken by the kernel just before calling ipoib_stop as part of devinet_ioctl
then it waits to all running tasks on ipoib_workqueue to complete. In case that "carrier on task" is running
it needs this lock to finish and machine gets into deadlock.

Signed-off-by: Yishai Hadas <yishaih@mellanox.com>
Signed-off-by: Erez Shitrit <erezsh@mellanox.com>