ib/ipoib: repalce list_del of the neigh->list with  list_del_init
In order to reslove situation where few process are trying to do
list_del(&neigh->list) and one of them release the spin_lock before it
tries again to tauch the list.
the fix contains 2 steps the first one is the replacement, that will
verify that the list can be deleted once.
(all the list_del_init are protected via spin_lock.)

RM #96472

Signed-off-by: Erez Shitrit <erezsh@mellanox.com>