mlx4_ib: Called rdma_port_get_link_layer with wrong port number.
rmda_port_get_link-layer was called with ports 0..1, instead of
1..2.
In IB mode, this caused the driver to think that port 1 was an ETH
port, resulting in registering a MAC address.  This invoked SET_PORT,
which set the port to ETH type. Since the port type was supposed to be
IB, port 1 remained disabled.

Fixes RM #120069

Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Yevgeny Petrilin <yevgenyp@mellanox.com>