net/mlx4: Add support to set default vlan id and priority for VF NIC
Add support for ndo_set_vf_vlan to the Ethernet driver.
Once invoked, this vport is considered to be in VST mode.
In this mode, the PPF driver configures Ethernet QPs created by this VF to
use the given vlan id and priority. Currently RoCE isn't supported in VST mode.

The special values of VID=4095 or VID=0,UP=0 are considered as VGT.

issue: 91903

Change-Id: I776a3843b24b01942d297820ce8ca1dd8844af79
Signed-off-by: Rony Efraim <ronye@mellanox.com>