net/mlx4_en: Fix bridged vSwitch configuration for non SRIOV mode
Commit 5b4c4d36860e "mlx4_en: Allow communication between functions on
same host" introduced a regression under which a bridge acting as vSwitch
whose uplink is an mlx4 Ethernet device become non-operative in native
(non sriov) mode. This happens since broadcast ARP requests sent by VMs
were loopback-ed by the HW and hence the bridge learned VM source MACs
on both the VM and the uplink ports.

The fix is to place the DMAC in the send WQE only under SRIOV/eSwitch
configuration or when the device is in selftest.

issue: 165570
issue: 130519

Change-Id: Ic203b8f65bd581dc16d6a854b711079a3cf179c6
Signed-off-by: Yan Burman <yanb@mellanox.com>