IB/mlx4: in RoCE, report network-view of number of GIDS in query_port.
eth_link_query_port() does not report the network-view number of GIDs (128)
when requested, so the gid table shown in sysfs under the iov directory shows
only the PF host view (which is 1 GID).

Fix this by returning 128 for the number of GIDs when the netw_view argument
flag is true.

issue: 131283

Change-Id: I1fe32cd9580b5aa61b63c3101fd1d3251f3dbd4e
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>