mlx4_core: fix reg/unreg vlan and reg/unreg mac to conform to the PRM
The functions mlx4_register_vlan, mlx4_unregister_vlan, mlx4_register_mac,
mlx4_unregister_mac make illegal use of the out_param in multifunc mode
to pass the port number. The PRM specifies that the port number should be
passed in bits 8..15 of the input-modifier field for ALLOC_RES and
FREE_RES (PRM 1.0 sections 20.15.1 and 20.15.2)

Since in old guests, the reg/unreg vlan in SRIOV mode was broken, old guests
will not work with the new code (vlan registration will simply fail).
For MAC register/unregister, a workaround is in place so that old guests will
continue to work on a new Hypervisor.

Note, though, that new guests will need to run over a new Hypervisor in order
to successfully register MACs and VLANS (because the high-order bits of the
input modifier will no longer be zero, but will contain the port number).

issue: 167687

Change-Id: I868c9e944ccd8bcf55b7397cec3cbbf6559d9691
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>