IB/mlx4: Do reg/unreg vlan for RoCE
The IB side of RoCE requires the true vlan index for the path
record in modify_qp.

We need to protect against the ETH side deleting or modifying
that VLAN table entry while the QP is active.

Solve this by registering for that vlan_id (thus obtaining the
vlan index, and also incrementing the ref-count for that vlan id).

issue: 130526

Change-Id: Id3375f43d71d516ad9dfefa222c795d318b43772
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>