ib/ipoib: Fix ipoib_neigh hashing to use the correct daddr octets
The hash function introduced in upstream commit b63b70d877 "IPoIB: Use a
private hash table for path lookup in xmit path" was designd to use
the 3 octets of the IPoIB HW address that holds the remote QPN. However,
this currently isn't the case under little endian machines as the code
there uses the flags part (octet[0]) and not the last octet of the QPN
(octet[3]), fix that.

The fix caused a checkpatch warning on line over 80 characters, to
solve that changed the name of the temp variable that holds the daddr.

issue: 169027

Change-Id: I70cf8c0e1580fde05f0a3f970854991df802f93a
Signed-off-by: Shlomo Pongratz <shlomop@mellanox.com>