mlx4: change QP allocation scheme
QP allocation change:

When using BF the QPN overrides the VLAN, CV, and SV fields in the WQE.
Thus, BF may only be used for QPNs with bits 6,7 unset.

Current EN code reserves a Tx QP range with 256b alignment.
This is a *bug* because if there are more than 64 Tx QPs in use
(will be a common case for 16-core machines),
because QPNs >= base + 65 will have set 6/7 bits.
Also, using ranges is not necessary here and is wasteful.

New mechanism will support reservation for "Eth QPs eligible for BF" for all drivers:
bare-metal, multi-PF, and eventually VFs (when hypervisors support WC in VMs).
Suggestions:
1.	In mlx4_en, allocate Tx QPs one by one instead of a range allocation,
	and request "BF enabled QPs" if BF is supported for the function
2.	In the ALLOC_RES PRM command, change param1 to:
a.	param1[23:0] - number of QPs
b.	param1[31] - Reservation for Eth QP supporting BF
c.	param1[30-24] - reserved
3.	In the wrapper for ALLOC_RES, if Eth QPs were requested,
	allocate only QPs with bits 6+7 clear

Note that the PRM change should not affect existing drivers
because the MSBs of param1 were always 0.

In the future we can add a "Eth BF QP supported" bit in QUERY_FUNC_CAP for this.
To make this a bit less drastic, we can set the "Eth BF QP" flag
only when BF is supported on guests. Currently, due to lack of KVM support
for write-combining on guests, BF is disabled for guests,
so old guest OSes can continue to work with the new hypervisor.
Old guests will not try to use the new allocation scheme.

We could couple BF allocation with BF enablement and save the cap bit,
but this assumes that VMs will not attempt BF QP allocation
if they weren't notified that BF is supported.

The final scheme:

New hypervisor and old guest:
Guest is not aware of new capability bit in QUERY_FUNC_CAP.
Hypervisor will get request to allocate aligned to 256 range of QPs.
Since BF on guests is not supported it will allocate the requested range.

Old hypervisor and new guest:
Guest will check BF capability bit:
- if it is 0 it will request for desired number of QPs according to the above
  ALLOC_RES format (e.g. bit 31 will be 0)
- it is not possible for capability bit to be 1 on old hypervisor

New hypervisor and new guest:
According to capability bit ask for desired number of QPs.

issue: 27033
issue: 165470

Change-Id: I57741c61d196c0e367d36817f98849da63090753
Signed-off-by: Eugenia Emantayev <eugenia@mellanox.co.il>